Renamed FacilitatorCommand to FacilitatorStart
